Rename Inpsyde\Helpers to Inpsyde\CodingStandard\Helpers
