+++ Inpsyde/Sniffs/CodeQuality/ArgumentTypeDeclarationSniff.php
@@ -28,9 +28,9 @@ declare(strict_types=1);
 
 namespace Inpsyde\Sniffs\CodeQuality;
 
-use Inpsyde\Helpers\FunctionDocBlock;
-use Inpsyde\Helpers\Functions;
-use Inpsyde\Helpers\WpHooks;
+use Inpsyde\CodingStandard\Helpers\FunctionDocBlock;
+use Inpsyde\CodingStandard\Helpers\Functions;
+use Inpsyde\CodingStandard\Helpers\WpHooks;
 use PHP_CodeSniffer\Sniffs\Sniff;
 use PHP_CodeSniffer\Files\File;
 use PHPCSUtils\Utils\FunctionDeclarations;
