+++ Inpsyde/Sniffs/CodeQuality/ReturnTypeDeclarationSniff.php
@@ -28,11 +28,11 @@ declare(strict_types=1);
 
 namespace Inpsyde\Sniffs\CodeQuality;
 
-use Inpsyde\Helpers\FunctionDocBlock;
-use Inpsyde\Helpers\FunctionReturnStatement;
-use Inpsyde\Helpers\Functions;
-use Inpsyde\Helpers\Misc;
-use Inpsyde\Helpers\WpHooks;
+use Inpsyde\CodingStandard\Helpers\FunctionDocBlock;
+use Inpsyde\CodingStandard\Helpers\FunctionReturnStatement;
+use Inpsyde\CodingStandard\Helpers\Functions;
+use Inpsyde\CodingStandard\Helpers\Misc;
+use Inpsyde\CodingStandard\Helpers\WpHooks;
 use PHP_CodeSniffer\Sniffs\Sniff;
 use PHP_CodeSniffer\Files\File;
 use PHPCSUtils\Utils\FunctionDeclarations;
