+++ composer.json
@@ -43,7 +43,7 @@
     },
     "autoload": {
         "psr-4": {
-            "Inpsyde\\Helpers\\": "Inpsyde/Helpers/"
+            "Inpsyde\\CodingStandard\\Helpers\\": "Inpsyde/Helpers/"
         }
     },
     "autoload-dev": {
