+++ tests/cases/Helpers/FunctionReturnStatementTest.php
@@ -29,7 +29,7 @@ declare(strict_types=1);
 namespace Inpsyde\CodingStandard\Tests\Helpers;
 
 use Inpsyde\CodingStandard\Tests\TestCase;
-use Inpsyde\Helpers\FunctionReturnStatement;
+use Inpsyde\CodingStandard\Helpers\FunctionReturnStatement;
 
 class FunctionReturnStatementTest extends TestCase
 {
